start conforming to norm
