+++ parser_obj_hittable.c
