+++ parser_obj_scene.c
