+++ parser_strtod.c
@@ -2,6 +2,7 @@
 #include "vec3d.h"
 #include "structs.h"
 #include "miniRT.h"
+#include "parse.h"
 
 static double	get_int_part(const char *s)
 {
